Change S1, S2 arrays from 'float' to 'unsigned int'.
When I ran the unit tests, I was initially surprised to see test failure for the downsample kernel. However, the apparent failure is just roundoff error because we use 'float' for the S1/S2 output arrays.

After thinking about it, I propose that we switch to using 'unsigned int' for the S1/S2 output arrays. It will make testing easier, and I didn't have a strong reason for making them 'float' in the first place!

This commit is a lot of lines of code, but it's just changing float -> uint in a lot of places. (The code changes were mostly done with search-and-replace :))

After this commit, the downsample() unit test passes -- the CPU and GPU kernels give bit-identical output.